Hitboxes hitboxes hitboxes

+++ Platformer/src/com/gravity/map/TileWorld.java
@@ -56,7 +56,7 @@ public class TileWorld implements GameWorld {
     public static final float STOMP_SPEED_FORWARD = 50.0f;
     public static final float STOMP_SPEED_BACKWARD = 30.0f;
 
-    public static final float TILE_MARGIN = 3f;
+    public static final int TILE_MARGIN = 3;
 
     public final int height;
     public final int width;
